put, delete, post, get

+++ routes/goods.js
@@ -1,6 +1,27 @@
 const express = require("express");
 const router = express.Router();
 const Goods = require("../schemas/goods");
+const Cart = require("../schemas/cart.js");
+const authMiddleware = require("../middlewares/auth-middleware");
+
+// Cart Reading API
+router.get("/goods/cart", authMiddleware, async (req, res) => {
+  const { userId } = res.locals.user;
+  const carts = await Cart.find({ userId }).exec();
+  const goodsIds = carts.map((cart) => {
+    return cart.goodsId;
+  });
+  const goods = await Goods.find({ goodsId: goodsIds });
+  const results = carts.map((cart) => {
+    return {
+      quantity: cart.quantity,
+      goods: goods.find((item) => item.goodsId === cart.goodsId),
+    };
+  });
+  res.json({
+    carts: results,
+  });
+});
 
 //상품 목록 조회 API
 router.get("/goods", async (req, res) => {
@@ -37,14 +58,13 @@ router.get("/goods/:goodsId", async (req, res) => {
 
   res.status(200).json({ goods: result });
 });
-
-const Cart = require("../schemas/cart.js");
-
-router.post("/goods/:goodsId/cart", async (req, res) => {
+// Post item to My cart API
+router.post("/goods/:goodsId/cart", authMiddleware, async (req, res) => {
+  const { userId } = res.locals.user;
   const { goodsId } = req.params;
   const { quantity } = req.body;
 
-  const existsCarts = await Cart.find({ goodsId });
+  const existsCarts = await Cart.find({ userId, goodsId });
   if (existsCarts.length) {
     return res.status(400).json({
       success: false,
@@ -52,31 +72,32 @@ router.post("/goods/:goodsId/cart", async (req, res) => {
     });
   }
 
-  await Cart.create({ goodsId, quantity });
+  await Cart.create({ userId, goodsId, quantity });
 
   res.json({ result: "success" });
 });
-
-router.put("/goods/:goodsId/cart", async (req, res) => {
+//My Cart Modifying PUT API
+router.put("/goods/:goodsId/cart", authMiddleware, async (req, res) => {
+  const { userId } = res.locals.user;
   const { goodsId } = req.params;
   const { quantity } = req.body;
 
-  const existsCarts = await Cart.find({ goodsId });
+  const existsCarts = await Cart.find({ userId, goodsId });
   if (existsCarts.length) {
     await Cart.updateOne(
-      { goodsId: goodsId },
+      { userId, goodsId: goodsId },
       { $set: { quantity: quantity } }
     );
   }
   res.status(200).json({ success: true });
 });
-
-router.delete("/goods/:goodsId/cart", async (req, res) => {
+//Cart Deleting
+router.delete("/goods/:goodsId/cart", authMiddleware, async (req, res) => {
   const { goodsId } = req.params;
-
+  const { userId } = res.locals.user;
   const existsCarts = await Cart.find({ goodsId });
   if (existsCarts.length) {
-    await Cart.deleteOne({ goodsId });
+    await Cart.deleteOne({ userId, goodsId });
   }
 
   res.json({ result: "success" });
